+++ KDSApiPadrao/KDS.Api/KDS.Api/Repositorio/Interface/IRepository.cs
@@ -14,8 +14,8 @@ namespace KDS.Api.Repositorio.Interface
         Pedido PegaItensPorPedido(Pedido pedido);
         IQueryable<Pedido> RetornaPedidos();
         Comanda InserePedido(Comanda comanda, string canaldeEntrada);
-        bool AlteraStatusPedido(Pedido pedido, int idStatus = 0);
-        bool AlteraStatusItem(int idPedido, int idItem, int idStatus);
+        bool AlteraStatusPedido(Pedido pedido, long idStatus = 0);
+        bool AlteraStatusItem(long idPedido, long idItem, long idStatus);
 
 
     }
